requirements.txt
flask>=3.0.0,<4.0
flask-cors>=4.0.0,<5.0
flask-limiter>=3.5.0,<4.0
scikit-learn>=1.4.0,<1.6
pandas>=2.0.0,<3.0
numpy>=1.26.0,<2.0
joblib>=1.3.0,<2.0
python-dotenv>=1.0.0,<2.0
gunicorn>=21.2.0,<23.0
PyJWT>=2.8.0,<3.0
bcrypt>=4.1.0,<5.0
pytest>=8.0.0
